Fix heredoc problems due to forking
When multiple lines are passed in as STDIN to minishell,
get_next_line with a default buffer size of 4096 grabs the entire
STDIN and stores it in a buffer. When we do the funny forking
for redirections, it clones over the buffer and we now have 2 copies
of "STDIN". When heredoc calls get_next_line on STDIN it is not
actually looking at the file offset, but grabbing off the line
from the buffer. After heredoc finishes, the parent continues
interpreting the next lines as input because it was unaffected
by the heredoc.